moved badges to subpage, wip on homepage
