+++ src/routes/investigation/+page.svelte
@@ -1,0 +1,18 @@
+<svelte:head>
+    <title>Ingress Plus &middot; Investigation</title>
+</svelte:head>
+
+<h2>The Investigation</h2>
+<h3>A complete story recap of Ingress</h3>
+
+<p>I will lose my mind before I complete this section. /.ixm</p>
+<p>I will also lose my mind if I hear one more "Investigation is mispelled" (Thanks for the heads up yall &lt;3) /.ixm</p>
+
+<style>
+    p {
+        text-align: center;
+        margin: 2em auto;
+        max-width: 800px;
+    }
+
+</style>
